+++ CHANGELOG.md
@@ -12,6 +12,7 @@
  - Add earth and stream spells (MIGRATION NOTE: lava spell needs updating in spells.yml, LavaSpell -> IterateSpell)
  - Add "/magic search" and "/magic cancel" commands.
  - Add effects to pillar, earth, stream, lava, bridge and stalactite spells.
+ - Add build permission check to the grenade, projectile and familiar spells.
 
 ## 2.8.2